[REFACTOR] Refactor File Location due to Inhencement Managements
