+++ app/src/main/java/com/example/pocky/presentation/screen/order/SandwitchActivity.java
@@ -1,4 +1,4 @@
-package com.example.pocky;
+package com.example.pocky.presentation.screen.order;
 
 import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
@@ -10,35 +10,41 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
+import com.example.pocky.R;
 
 import java.util.Arrays;
 import java.util.List;
 
-public class SmilesuppActivity extends AppCompatActivity {
+public class SandwitchActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_smilesupp);
+        setContentView(R.layout.activity_sandwitch);
 
         RecyclerView recyclerView = findViewById(R.id.recyclerView);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
         List<Integer> imageList = Arrays.asList(
-                R.drawable.resize_baconcheesewedgepotato,
-                R.drawable.resize_cheesewedgepotato,
-                R.drawable.resize_chickenbaconminiwrap,
-                R.drawable.resize_chips,
-                R.drawable.resize_chocolatechip,
-                R.drawable.resize_cornsoup,
-                R.drawable.resize_doublechocolatechip,
-                R.drawable.resize_hashbrown,
-                R.drawable.resize_milk,
-                R.drawable.resize_mushroomsoup,
-                R.drawable.resize_oatmealrasin,
-                R.drawable.resize_raspberrycheesecake,
-                R.drawable.resize_wedgepotato,
-                R.drawable.resize_whitechocomacadamia
+                R.drawable.resize_bltsandwitch,
+                R.drawable.resize_chickenbaconavocadosandwitch,
+                R.drawable.resize_chickenslicesandwitch,
+                R.drawable.resize_chickenteriyakisandwitch,
+                R.drawable.resize_eggslicesandwitch,
+                R.drawable.resize_eggmayosandwitch,
+                R.drawable.resize_hamsandwitch,
+                R.drawable.resize_italianbmtsandwitch,
+                R.drawable.resize_kbbqsandwitch,
+                R.drawable.resize_pullporkcheesesandwitch,
+                R.drawable.resize_roastedchickensandwitch,
+                R.drawable.resize_rotisseriebbqchickensandwitch,
+                R.drawable.resize_shrimpsandwitch,
+                R.drawable.resize_spicyitaliansandwitch,
+                R.drawable.resize_spicyshrimpsandwitch,
+                R.drawable.resize_steakandcheesesandwitch,
+                R.drawable.resize_steakcheesesalad,
+                R.drawable.resize_subwayclubsandwitch,
+                R.drawable.resize_veggiesandwitch
         );
 
         BreadAdapter adapter = new BreadAdapter(imageList);
