+++ app/src/main/java/com/example/pocky/presentation/screen/order/SmilesuppActivity.java
@@ -1,4 +1,4 @@
-package com.example.pocky;
+package com.example.pocky.presentation.screen.order;
 
 import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
@@ -10,24 +10,36 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
+import com.example.pocky.R;
 
 import java.util.Arrays;
 import java.util.List;
 
-public class WrappActivity extends AppCompatActivity {
+public class SmilesuppActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_wrapp);
+        setContentView(R.layout.activity_smilesupp);
 
         RecyclerView recyclerView = findViewById(R.id.recyclerView);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
         List<Integer> imageList = Arrays.asList(
+                R.drawable.resize_baconcheesewedgepotato,
+                R.drawable.resize_cheesewedgepotato,
                 R.drawable.resize_chickenbaconminiwrap,
-                R.drawable.resize_shrimpeggmayorap,
-                R.drawable.resize_steakncheeseavocadowrap
+                R.drawable.resize_chips,
+                R.drawable.resize_chocolatechip,
+                R.drawable.resize_cornsoup,
+                R.drawable.resize_doublechocolatechip,
+                R.drawable.resize_hashbrown,
+                R.drawable.resize_milk,
+                R.drawable.resize_mushroomsoup,
+                R.drawable.resize_oatmealrasin,
+                R.drawable.resize_raspberrycheesecake,
+                R.drawable.resize_wedgepotato,
+                R.drawable.resize_whitechocomacadamia
         );
 
         BreadAdapter adapter = new BreadAdapter(imageList);
